# src/api/specs/swagger.yaml
openapi: 3.0.0
info:
  title: Next.js API
  description: API documentation for Next.js 15 application
  version: 1.0.0
servers:
  - url: http://localhost:3000
    description: Local server
components:
  securitySchemes:
    jwt:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    NotFoundError:
      type: object
      properties:
        error:
          type: string
          example: NotFound
        message:
          type: string
          example: The target data does not exist.
...
    ValidationError:
      type: object
      properties:
        error:
          type: string
          example: ValidationError
        errors:
          type: array
          items:
            type: object
            properties:
              path:
                type: string
                example: email
              message:
                type: string
                example: Invalid email address
    InternalServerError:
      type: object
      properties:
        error:
          type: string
          example: InternalServerError
        message:
          type: string
          example: Server error.
    Success:
      type: object
      required:
        - message
      properties:
        message:
          type: string
          example: Success
...
paths:
...
  /api/admin/users/delete-user:
    delete:
      summary: Delete users by IDs
      description: >-
        Admin can delete user accounts including their own, but not other
        admins.
      operationId: adminDeleteUsers
      tags:
        - Admin
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - ids
              properties:
                ids:
                  type: array
                  items:
                    type: string
                    format: uuid
                  example:
                    - 123e4567-e89b-12d3-a456-426614174000
                    - 123e4567-e89b-12d3-a456-426614174001
      responses:
        '200':
          description: Users deleted successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Success'
        '400':
          description: Validation errors or malformed request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationError'
        '404':
          description: Some users not found with the provided IDs
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NotFoundError'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InternalServerError'
  /api/admin/users/get-user:
    post:
      summary: Fetch users with pagination and search
      description: Retrieves users from the database with pagination and optional search.
      operationId: adminGetUsers
      tags:
        - Admin
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                page:
                  type: integer
                  example: 1
                limit:
                  type: integer
                  example: 10
                search:
                  type: string
                  example: john
      responses:
        '200':
          description: Successfully retrieved users with pagination
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      type: object
                      properties:
                        id:
                          type: string
                          format: uuid
                        email:
                          type: string
                        firstName:
                          type: string
                        lastName:
                          type: string
                        gender:
                          type: string
                        country:
                          type: string
                        address:
                          type: string
                        role:
                          type: string
                  pagination:
                    type: object
                    properties:
                      page:
                        type: integer
                        example: 1
                      limit:
                        type: integer
                        example: 10
                      showPerPage:
                        type: integer
                        example: 10
                      totalUsers:
                        type: integer
                        example: 50
                      totalPages:
                        type: integer
                        example: 5
        '400':
          description: Validation errors or malformed request
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: Validation error
                  errors:
                    type: array
                    items:
                      type: object
                      properties:
                        path:
                          type: string
                          example: page
                        message:
                          type: string
                          example: Page must be a number
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InternalServerError'
  /api/admin/users/update-user:
    put:
      summary: Update user details
      description: Update details of an existing user based on the provided user ID.
      operationId: adminUpdateUser
      tags:
        - Admin
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - id
              properties:
                id:
                  type: string
                  format: uuid
                  description: The ID of the user to update.
                  example: 123e4567-e89b-12d3-a456-426614174000
                email:
                  type: string
                  description: The email address of the user.
                  example: johndoe@example.com
                firstName:
                  type: string
                  description: The first name of the user.
                  example: John
                lastName:
                  type: string
                  description: The last name of the user.
                  example: Doe
                gender:
                  type: string
                  description: The gender of the user.
                  example: male
                address:
                  type: string
                  description: The address of the user.
                  example: 123 Main St, Springfield, IL
                role:
                  type: string
                  enum:
                    - user
                    - admin
                  description: The role of the user.
                  example: user
                country:
                  type: string
                  description: The country of the user.
                  example: USA
      responses:
        '200':
          description: User updated successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: User updated successfully
                  user:
                    type: object
                    properties:
                      id:
                        type: string
                        format: uuid
                        example: 123e4567-e89b-12d3-a456-426614174000
                      email:
                        type: string
                        example: johndoe@example.com
                      firstName:
                        type: string
                        example: John
                      lastName:
                        type: string
                        example: Doe
                      gender:
                        type: string
                        example: male
                      address:
                        type: string
                        example: 123 Main St, Springfield, IL
                      role:
                        type: string
                        enum:
                          - USER
                          - ADMIN
                        example: USER
                      country:
                        type: string
                        example: USA
        '400':
          description: Validation errors or malformed request
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: Validation error
                  errors:
                    type: array
                    items:
                      type: object
                      properties:
                        path:
                          type: string
                          example: email
                        message:
                          type: string
                          example: Invalid email format
        '404':
          description: User not found
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: User not found
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: Internal server error
                  message:
                    type: string
                    example: Error during registration
...